requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
coverage==4.5.1
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gevent==1.1.2
greenlet==0.4.14
grpcio==1.13.0
gunicorn==19.3.0
gunicorn-thrift==0.2.21
more-itertools==4.2.0
pluggy==0.6.0
ply==3.11
protobuf==3.6.0
py==1.5.4
pytest==3.6.3
pytest-cov==2.5.1
setproctitle==1.1.10
six==1.11.0
thrift==0.9.3
thriftpy==0.3.e
